Add constraint to ensure a maxTimeInVehicle for new shipments
Make use of a MaximumTimeInVehicle HardActivityConstraint:
Unfortunately the activities themself don't update their end and arrivalTimes,
but the real depatureTime of the previous activity will be passed to the method
and we can make use of that to calculate the time spent in a vehicle.

The calculation itself is simple, yet there are different scenarios which
needs to be considered, e.g. the new shipment could be part of a route already,
it could be planned to be part of the route or it could be totally new.

A StateManager is used to keep track of temporary states and helps recalculating
the maximumTimeInVehicle whenenver a newAct will be inserted in the route where
this constraint is present for its shipment.